+++ src/web/GliderTable.cpp
@@ -275,7 +275,7 @@ void GliderTable::loadPage(unsigned int page)
 
     clear();
     createHeaderRow();
-    const int nFirst = (pageNr_ - 1) * entriesPerPage_;
+    const size_t nFirst = (pageNr_ - 1) * entriesPerPage_;
     if(nFirst < gliders_.size())
     {
         vector<shared_ptr<Glider> >::iterator ibeg = gliders_.begin();
